refactor: move config parsing to Radio
