+++ examples/ping-pong/src/config.rs
@@ -1,0 +1,76 @@
+use clap::Parser;
+use ethers::signers::WalletError;
+use graphcast_sdk::build_wallet;
+use graphcast_sdk::graphcast_id_address;
+use graphcast_sdk::init_tracing;
+use serde::{Deserialize, Serialize};
+use tracing::info;
+
+#[derive(Clone, Debug, Parser, Serialize, Deserialize)]
+#[clap(
+    name = "poi-radio",
+    about = "Cross-check POIs with other Indexer in real time",
+    author = "GraphOps"
+)]
+pub struct Config {
+    #[clap(
+        long,
+        value_name = "ENDPOINT",
+        env = "GRAPH_NODE_STATUS_ENDPOINT",
+        help = "API endpoint to the Graph Node Status Endpoint"
+    )]
+    pub graph_node_endpoint: String,
+    #[clap(
+        long,
+        value_name = "KEY",
+        value_parser = Config::parse_key,
+        env = "PRIVATE_KEY",
+        hide_env_values = true,
+        help = "Private key to the Graphcast ID wallet (Precendence over mnemonics)",
+    )]
+    pub private_key: Option<String>,
+    #[clap(
+        long,
+        value_name = "KEY",
+        value_parser = Config::parse_key,
+        env = "MNEMONIC",
+        hide_env_values = true,
+        help = "Mnemonic to the Graphcast ID wallet (first address of the wallet is used; Only one of private key or mnemonic is needed)",
+    )]
+    pub mnemonic: Option<String>,
+    #[clap(
+        long,
+        value_name = "SUBGRAPH",
+        env = "REGISTRY_SUBGRAPH",
+        help = "Subgraph endpoint to the Graphcast Registry",
+        default_value = "https://api.thegraph.com/subgraphs/name/hopeyen/graphcast-registry-goerli"
+    )]
+    pub registry_subgraph: String,
+    #[clap(
+        long,
+        value_name = "SUBGRAPH",
+        env = "NETWORK_SUBGRAPH",
+        help = "Subgraph endpoint to The Graph network subgraph",
+        default_value = "https://gateway.testnet.thegraph.com/network"
+    )]
+    pub network_subgraph: String,
+}
+
+impl Config {
+    /// Parse config arguments
+    pub fn args() -> Self {
+        // TODO: load config file before parse (maybe add new level of subcommands)
+        let config = Config::parse();
+        init_tracing().expect("Could not set up global default subscriber for logger, check environmental variable `RUST_LOG` or the CLI input `log-level`");
+        config
+    }
+
+    /// Validate that private key as an Eth wallet
+    fn parse_key(value: &str) -> Result<String, WalletError> {
+        // The wallet can be stored instead of the original private key
+        let wallet = build_wallet(value)?;
+        let addr = graphcast_id_address(&wallet);
+        info!("Resolved Graphcast id: {}", addr);
+        Ok(String::from(value))
+    }
+}
