+++ src/lib.rs
@@ -41,7 +41,6 @@ use url::{Host, Url};
 use waku::WakuPubSubTopic;
 
 pub mod bots;
-pub mod config;
 pub mod graphcast_agent;
 pub mod graphql;
 pub mod networks;
